#587 Server - Api: Song Handler - Migrate Section Routes to separate handler, api requests and router + New [PUT] Endpoint: Bulk Delete Song Sections + Testing
