+++ repertoire.server/domain/usecase/song/section/delete_song_section.go
@@ -38,6 +38,7 @@ func (d DeleteSongSection) Handle(id uuid.UUID, songID uuid.UUID) *wrapper.Error
 		return wrapper.NotFoundError(errors.New("song section not found"))
 	}
 
+	// reorder the other sections
 	sectionsLength := len(song.Sections)
 	for i := index + 1; i < sectionsLength; i++ {
 		song.Sections[i].Order = song.Sections[i].Order - 1
@@ -58,7 +59,7 @@ func (d DeleteSongSection) Handle(id uuid.UUID, songID uuid.UUID) *wrapper.Error
 	if err != nil {
 		return wrapper.InternalServerError(err)
 	}
-	err = d.songRepository.DeleteSection(id)
+	err = d.songRepository.DeleteSections([]uuid.UUID{id})
 	if err != nil {
 		return wrapper.InternalServerError(err)
 	}
